refactor form
