+++ playground/src/components/form/index.tsx
@@ -1,24 +1,9 @@
-import { RJSFSchema, toPathSchema, UIOptionsType, ValidatorType, PathSchema } from '@rjsf/utils';
-import Form from '@rjsf/core';
+import { ThemeProps, withTheme } from "@rjsf/core";
+import TextAreaWidget from "./textarea_widget";
 
+const theme: ThemeProps = {
+  widgets: { textarea: TextAreaWidget },
+};
 
-export function getUISchema(validator: ValidatorType, schema: RJSFSchema, value: (property: string) => UIOptionsType) {
-  const paths = toPathSchema(validator, schema, undefined, schema);
-  function recursiveSetValueForPaths(property: string, paths: PathSchema) {
-    const properties = Object.keys(paths).filter((path) => path !== "$name");
-    if (properties.length === 0) {
-      return value(property);
-    }
-    return {
-      ...properties.reduce((acc, property) => {
-        acc[property] = recursiveSetValueForPaths(property, paths[property]); 
-        return acc;
-      }
-      , {}),
-    }
-  }
-  return recursiveSetValueForPaths("__root__", paths);
-}
-  
-
-export default Form;
+const ThemedForm = withTheme(theme);
+export default ThemedForm;
