+++ playground/src/components/module_config/module_llm.tsx
@@ -1,65 +1,23 @@
 import { useState } from "react";
 
-import { UIOptionsType, getDefaultFormState, WidgetProps } from "@rjsf/utils";
+import { UIOptionsType, getDefaultFormState, WidgetProps, DescriptionFieldProps, RegistryFieldsType } from "@rjsf/utils";
 import validator from "@rjsf/validator-ajv8";
 import { Editor, Monaco } from "@monaco-editor/react";
 import { editor } from "monaco-editor";
 
-import Form, { getUISchema } from "@/components/form";
+import Form from "@/components/form";
+import { getUISchema } from "@/components/form/utils";
 import { ModuleConfigProps } from ".";
 
-const PromptTextAreaWidget = ({ id, value, onChange }: WidgetProps) => {
-  const [height, setHeight] = useState(100);
 
-  const handleEditorDidMount = (
-    editor: editor.IStandaloneCodeEditor,
-    monaco: Monaco
-  ) => {
-    editor.onDidContentSizeChange(() => {
-      setHeight(Math.min(1000, editor.getContentHeight()));
-      editor.layout();
-    });
-
-    monaco.languages.register({ id: "placeholder" });
-    monaco.languages.setMonarchTokensProvider("placeholder", {
-      tokenizer: {
-        root: [[/{\w*}/, "variable"]],
-      },
-    });
-    monaco.editor.defineTheme("my-theme", {
-      base: "vs",
-      inherit: true,
-      rules: [{ token: "variable", foreground: "0000FF" }],
-      colors: {},
-    });
-    monaco.editor.setTheme("my-theme");
-  };
-
-  return (
-    <div id={id} className="border border-gray-300 rounded my-2 py-2">
-      <div style={{ height }}>
-        <Editor
-          options={{
-            scrollBeyondLastLine: false,
-            automaticLayout: true,
-            scrollbar: {
-              vertical: "hidden",
-              horizontal: "hidden",
-              alwaysConsumeMouseWheel: false,
-            },
-            minimap: {
-              enabled: false,
-            },
-          }}
-          value={value}
-          onMount={handleEditorDidMount}
-          defaultLanguage="placeholder"
-          onChange={(value) => onChange(value)}
-        />
-      </div>
-    </div>
-  );
+const CustomDescriptionField = ({id, description}) => {
+  return <div id={id}>TEST: {description}</div>;
 };
+
+const fields = {
+  DescriptionField: CustomDescriptionField
+};
+
 
 export default function ModuleLLMConfig({
   moduleConfig,
@@ -77,6 +35,7 @@ export default function ModuleLLMConfig({
         formData={moduleConfig}
         liveValidate
         className="schema-form"
+        fields={fields}
         uiSchema={{
           "ui:submitButtonOptions": {
             norender: true,
@@ -87,7 +46,7 @@ export default function ModuleLLMConfig({
               "ui:enableMarkdownInDescription": true,
             };
             if (property.includes("message")) {
-              config["ui:widget"] = PromptTextAreaWidget;
+              config["ui:widget"] = "textarea";
             }
             return config;
           }),
